# docs/swagger.yaml
definitions:
  api.balitaResponse:
    properties:
      berat_lahir:
        type: string
      created_date:
        type: string
      id:
        type: string
      id_keluarga:
        type: string
      jenis_kelamin:
        type: string
      kecamatan:
        type: string
      kelurahan:
        type: string
      nama:
        type: string
      nama_ayah:
        type: string
      nama_ibu:
        type: string
      nomor_kk:
        type: string
      tanggal_lahir:
        type: string
      tinggi_lahir:
        type: string
      umur:
        description: calculated field in months
        type: string
      updated_date:
        type: string
    type: object
...
  api.getAllBalitaResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/api.balitaResponse'
        type: array
      total:
        type: integer
    type: object
...
  api.insertBalitaRequest:
    properties:
      berat_lahir:
        description: in grams
        type: string
      id_keluarga:
        type: string
      jenis_kelamin:
        description: '"L" or "P"'
        type: string
      nama:
        type: string
      tanggal_lahir:
        description: 'Format: YYYY-MM-DD'
        type: string
      tinggi_lahir:
        description: in cm
        type: string
    type: object
  api.insertBalitaResponse:
    properties:
      id:
        type: string
    type: object
...
  object.Response:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
    url: http://www.swagger.io/support
  description: API for managing stunting data
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Stunting Web API
  version: 0.0.2
paths:
...
  /api/admin/balita/get:
    get:
      consumes:
      - application/json
      description: |-
        Get balita data based on query parameter (Admin only)

        Response data varies by parameter:
        - Without id parameter: Returns all balita with total count
        - With id parameter: Returns specific balita data

        Balita data includes: nama, tanggal_lahir, jenis_kelamin, berat_lahir, tinggi_lahir, umur, keluarga info, location info
      parameters:
      - description: Balita ID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Balita data retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.getAllBalitaResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Balita not found
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - Bearer: []
      summary: Get balita data
      tags:
      - admin
  /api/admin/balita/insert:
    post:
      consumes:
      - application/json
      description: |-
        Insert new balita data (Admin only)

        Inserts balita record with data including:
        - id_keluarga, nama, tanggal_lahir, jenis_kelamin
        - berat_lahir (in grams), tinggi_lahir (in cm)
        - Validates keluarga existence and balita age criteria (under 5 years)
      parameters:
      - description: Balita data
        in: body
        name: balita
        required: true
        schema:
          $ref: '#/definitions/api.insertBalitaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Balita inserted successfully
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.insertBalitaResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/object.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - Bearer: []
      summary: Insert new balita
      tags:
      - admin
...
swagger: "2.0"